Ems 34 implement parser to handle ssh output

+++ src/main/java/vn/com/tma/emsbackend/dto/NetworkDeviceRequestDto.java
@@ -17,7 +17,7 @@ public class NetworkDeviceRequestDto {
     private String ipAddress;
 
     @Port(message = "Invalid port number")
-    private long port;
+    private int port;
 
     @NotNull(message = "Credential can not empty")
     private long credentialId;
